+++ src/shared/home/my-votes.tsx
@@ -17,14 +17,9 @@ import { MyVotesTable } from "./my-votes-table";
 
 type State = {
   showVoteNowModal: boolean;
-  isBlur: boolean;
 };
 
-type Props = {
-  isMobile?: boolean;
-  isIoPayMobile?: boolean;
-  isInAppWebview?: boolean;
-};
+type Props = {};
 
 // path: /my-votes
 export function MyVotes(): JSX.Element {
@@ -37,21 +32,11 @@ export function MyVotes(): JSX.Element {
   );
 }
 
-// @ts-ignore
-@connect((state) => ({
-  // @ts-ignore
-  isIoPayMobile: state.base.isIoPayMobile,
-  // @ts-ignore
-  isInAppWebview: state.base.isInAppWebview,
-  // @ts-ignore
-  isMobile: state.base.isMobile,
-}))
 export class StakingContractContainer extends PureComponent<Props, State> {
   constructor(props: Props) {
     super(props);
     this.state = {
       showVoteNowModal: false,
-      isBlur: false,
     };
   }
 
@@ -59,65 +44,20 @@ export class StakingContractContainer extends PureComponent<Props, State> {
     window.location.reload();
   };
 
-  componentDidMount(): void {
-    window.onblur = () => {
-      this.setState({
-        isBlur: true,
-      });
-    };
-  }
-  openDeepLink = () => {
-    const a = document.createElement("a");
-    const tagId = "startIoPay";
-    a.setAttribute("href", "iopay://io.iotex.iopay/open?action=stake");
-    a.setAttribute("id", tagId);
-    if (document.getElementById(tagId)) {
-      // @ts-ignore
-      document.body.removeChild(document.getElementById(tagId));
-    }
-    document.body.appendChild(a);
-    a.click();
-  };
-
-  ioPayIsInstall = () => {
-    this.openDeepLink();
-    setTimeout(() => {
-      if (!this.state.isBlur) {
-        location.href = "http://iopay.iotex.io";
-      }
-    }, 3000);
-  };
-
   render(): JSX.Element {
-    const { isIoPayMobile, isMobile } = this.props;
     return (
       <div>
         <SmartContractCalled />
-        {isMobile && !isIoPayMobile ? (
-          <VotingButton
-            launch={() => {
-              this.ioPayIsInstall();
-            }}
-            disabled={false}
-            extra={{ size: "large" }}
-          >
-            <span>
-              <PlusOutlined />
-              {t("my_stake.button.vote_with_iopay")}
-            </span>
-          </VotingButton>
-        ) : (
-          <VotingButton
-            launch={() => this.setState({ showVoteNowModal: true })}
-            disabled={false}
-            extra={{ size: "large" }}
-          >
-            <span>
-              <PlusOutlined />
-              {t("my_stake.new_vote")}
-            </span>
-          </VotingButton>
-        )}
+        <VotingButton
+          launch={() => this.setState({ showVoteNowModal: true })}
+          disabled={false}
+          extra={{ size: "large" }}
+        >
+          <span>
+            <PlusOutlined />
+            {t("my_stake.new_vote")}
+          </span>
+        </VotingButton>
 
         <RefreshButtonStyle onClick={() => this.refreshPage()}>
           <RedoOutlined style={{ marginRight: 4 }} />
